requirements.txt
autopep8==2.0.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
msgpack==1.0.4
numpy==1.23.4
pandas==1.5.1
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
autopep8==2.0.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
msgpack==1.0.4
numpy==1.23.4
oauthlib==3.2.2
pandas==1.5.1
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
